Move subscription stuff to own file
